project/apps/version_control/serializers/repositorie.py
from rest_framework import serializers
from .. import models


class RepositorieSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Repositorie
        fields = (
            'id', 'name'
        )